move teamCard one up
