+++ Frontend/src/components/teams/TeamCard/teamCard.module.css
